Move helpers out of XFile class
